Refactor of dead lines and additional content, addes extra content to scheme, added checking roles in web.api for this controllers

+++ src/Web.Api/Controllers/AdditionalContentPublicationsController.cs
@@ -1,44 +1,56 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.Serialization;
 using System.Threading.Tasks;
 using Database;
 using Database.Models;
 using Database.Repos;
 using Database.Repos.Groups;
 using Database.Repos.Users;
-using JetBrains.Annotations;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using Ulearn.Common.Api.Models.Responses;
 using Ulearn.Core.Courses.Manager;
-using Ulearn.Web.Api.Models.Common;
+using Ulearn.Web.Api.Models.Responses.AdditionalContent;
 
 namespace Ulearn.Web.Api.Controllers
 {
-	[Route("/additional-content")]
+	[Route("/additional-content-publications")]
+	[Authorize]
 	public class AdditionalContentPublicationsController : BaseController
 	{
 		private IAdditionalContentPublicationsRepo additionalContentPublicationsRepo;
+		private ICourseRolesRepo courseRolesRepo;
 		private IGroupsRepo groupsRepo;
 
 		public AdditionalContentPublicationsController(ICourseStorage courseStorage, UlearnDb db,
 			IUsersRepo usersRepo,
 			IGroupsRepo groupsRepo,
+			ICourseRolesRepo courseRolesRepo,
 			IAdditionalContentPublicationsRepo additionalContentPublicationsRepo
 		)
 			: base(courseStorage, db, usersRepo)
 		{
 			this.groupsRepo = groupsRepo;
+			this.courseRolesRepo = courseRolesRepo;
 			this.additionalContentPublicationsRepo = additionalContentPublicationsRepo;
 		}
 
 		[HttpGet]
-		[Route("{courseId}")]
-		public async Task<ActionResult<AdditionalContentPublicationsResponse>> GetAdditionalContentPublications([FromRoute] string courseId, [FromQuery] int groupId)
+		public async Task<ActionResult<AdditionalContentPublicationsResponse>> GetAdditionalContentPublications([FromQuery] string courseId, [FromQuery] int groupId)
 		{
+			var isTester = await courseRolesRepo.HasUserAccessToCourse(UserId, courseId, CourseRoleType.Tester).ConfigureAwait(false);
+
+			if (!isTester)
+			{
+				return Forbid($"You have no access to course {courseId}");
+			}
+			
+			var group = (await groupsRepo.GetCourseGroupsQueryable(courseId).Where(g => g.Id == groupId).ToListAsync()).FirstOrDefault();
+			if (group == null)
+			{
+				return NotFound($"Group with id {groupId} not found");
+			}
+			
 			var publications = await additionalContentPublicationsRepo.GetAdditionalContentPublications(courseId, groupId);
 			var userIds = publications.Select(p => p.AuthorId).Distinct().ToList();
 			var users = (await usersRepo.GetUsersByIds(userIds)).ToDictionary(u => u.Id, u => u);
@@ -52,9 +64,8 @@ namespace Ulearn.Web.Api.Controllers
 		}
 
 		[HttpPost]
-		[Route("{courseId}")]
 		public async Task<ActionResult<AdditionalContentPublicationResponse>> AddAdditionalContentPublication(
-			[FromRoute] string courseId,
+			[FromQuery] string courseId,
 			[FromQuery] int groupId,
 			[FromQuery] Guid unitId,
 			[FromQuery] Guid? slideId,
@@ -63,36 +74,31 @@ namespace Ulearn.Web.Api.Controllers
 		{
 			var course = courseStorage.FindCourse(courseId);
 			if (course == null)
-			{
-				return NotFound();
-			}
+				return NotFound($"Course {courseId} not found");
+			
+			var isTester = await courseRolesRepo.HasUserAccessToCourse(UserId, courseId, CourseRoleType.Tester).ConfigureAwait(false);
+
+			if (!isTester)
+				return Forbid($"You have no access to course {courseId}");
 
 			var unit = course.GetUnits(new[] { unitId }).FirstOrDefault();
 			if (unit == null)
-			{
-				return NotFound();
-			}
+				return NotFound($"Unit with id {unitId} not found");
 
 			if (slideId != null)
 			{
 				var slides = unit.GetSlides(false);
 				if (slides.All(s => s.Id != slideId))
-				{
-					return NotFound();
-				}
+					return NotFound($"Slide with id {slideId} not found");
 			}
-			
+
 			var group = (await groupsRepo.GetCourseGroupsQueryable(courseId).Where(g => g.Id == groupId).ToListAsync()).FirstOrDefault();
 
 			if (group == null)
-			{
-				return NotFound();
-			}
+				return NotFound($"Group with id {groupId} not found");
 
 			if (await additionalContentPublicationsRepo.HasPublication(courseId, groupId, unitId, slideId))
-			{
 				return UnprocessableEntity($"There's already a publication for this content, try update it instead");
-			}
 
 			var publication = await additionalContentPublicationsRepo.AddAdditionalContentPublication(courseId, groupId, UserId, unitId, slideId, date);
 			var user = (await usersRepo.GetUsersByIds(new[] { UserId })).FirstOrDefault();
@@ -101,14 +107,22 @@ namespace Ulearn.Web.Api.Controllers
 		}
 
 		[HttpPatch]
-		public async Task<ActionResult> UpdateAdditionalContentPublication([FromQuery] Guid publicationId, [FromQuery] DateTime date)
+		[Route("{publicationId}")]
+		public async Task<ActionResult> UpdateAdditionalContentPublication([FromRoute] Guid publicationId, [FromQuery] DateTime date)
 		{
 			var publicationToUpdate = await additionalContentPublicationsRepo.GetAdditionalContentPublicationById(publicationId);
 			if (publicationToUpdate == null)
-			{
 				return NotFound($"Publication with id {publicationId} not found");
-			}
+			
+			var course = courseStorage.FindCourse(publicationToUpdate.CourseId);
+			if (course == null)
+				return NotFound($"Course {publicationToUpdate.CourseId} not found");
+			
+			var isTester = await courseRolesRepo.HasUserAccessToCourse(UserId, publicationToUpdate.CourseId, CourseRoleType.Tester).ConfigureAwait(false);
 
+			if (!isTester)
+				return Forbid($"You have no access to course {publicationToUpdate.CourseId}");
+			
 			publicationToUpdate.Date = date;
 
 			await additionalContentPublicationsRepo.UpdateAdditionalContentPublication(publicationToUpdate);
@@ -117,64 +131,27 @@ namespace Ulearn.Web.Api.Controllers
 		}
 
 		[HttpDelete]
-		public async Task<ActionResult> DeleteAdditionalContentPublication([FromQuery] Guid publicationId)
+		[Route("{publicationId}")]
+		public async Task<ActionResult> DeleteAdditionalContentPublication([FromRoute] Guid publicationId)
 		{
 			var publication = await additionalContentPublicationsRepo.GetAdditionalContentPublicationById(publicationId);
 			if (publication == null)
 			{
 				return NotFound($"Publication with id {publicationId} not found");
 			}
+			
+			var course = courseStorage.FindCourse(publication.CourseId);
+			if (course == null)
+				return NotFound($"Course {publication.CourseId} not found");
+			
+			var isTester = await courseRolesRepo.HasUserAccessToCourse(UserId, publication.CourseId, CourseRoleType.Tester).ConfigureAwait(false);
+
+			if (!isTester)
+				return Forbid($"You have no access to course {publication.CourseId}");
 
 			await additionalContentPublicationsRepo.DeleteAdditionalContentPublication(publication);
 
 			return Ok($"Publication with id {publicationId} deleted");
-		}
-	}
-
-	[DataContract]
-	public class AdditionalContentPublicationsResponse
-	{
-		[DataMember]
-		public List<AdditionalContentPublicationResponse> Publications;
-	}
-
-	[DataContract]
-	public class AdditionalContentPublicationResponse
-	{
-		[DataMember]
-		public Guid Id;
-
-		[DataMember]
-		public string CourseId;
-
-		[DataMember]
-		public int GroupId;
-
-		[DataMember]
-		public Guid UnitId;
-
-		[DataMember]
-		[CanBeNull]
-		public Guid? SlideId;
-
-		[DataMember]
-		public DateTime Date;
-
-		[DataMember]
-		public ShortUserInfo Author;
-
-		public static AdditionalContentPublicationResponse Build(AdditionalContentPublication p, ApplicationUser user)
-		{
-			return new AdditionalContentPublicationResponse
-			{
-				Id = p.Id,
-				Author = BaseController.BuildShortUserInfo(user),
-				Date = p.Date,
-				UnitId = p.UnitId,
-				SlideId = p.SlideId,
-				CourseId = p.CourseId,
-				GroupId = p.GroupId,
-			};
 		}
 	}
 }
